requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
upstash-redis==1.5.0
uvicorn==0.40.0
